% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlackJack-package.R
\docType{package}
\name{BlackJack-package}
\alias{BlackJack}
\alias{BlackJack-package}
\title{BlackJack: Play BlackJack}
\description{
The purpose of my package is to play a game of blackjack.
}
\keyword{internal}
