% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlackJackDeal.R
\name{BlackJackDeal}
\alias{BlackJackDeal}
\title{Prepare the game}
\usage{
BlackJackDeal(Deck)
}
\arguments{
\item{Deck}{Enter the current deck within the function}
}
\value{
Creates a deck within the local environment, in order to make the needed changes
}
\description{
Prepare the game
}
\examples{
BlackJackDeal(Build_playing_deck(2))
}
