% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDeck.R
\name{Build_playing_deck}
\alias{Build_playing_deck}
\title{Build a Deck for this game}
\usage{
Build_playing_deck(x)
}
\arguments{
\item{x}{The number of decks in total contributing to the deck used in this game.}
}
\value{
This function will return our deck we decide to play with in the game.
}
\description{
Build a Deck for this game
}
\examples{
Build_playing_deck(2)
Build_playing_deck(3)
Build_playing_deck(4)
}
