% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{count_incrementer}
\alias{count_incrementer}
\title{Get value of card to update the count}
\usage{
count_incrementer(card)
}
\arguments{
\item{card}{A list containing two attributes: rank and suit}
}
\value{
1, if the rank is 10, J, Q, K, A. -1, if the rank is less than 7. Otherwise, 0.
}
\description{
Get value of card to update the count
}
