% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{create_game}
\alias{create_game}
\title{Create a blackjack game}
\usage{
create_game(chip_stacks, betting_units, minimum_bet, num_decks)
}
\arguments{
\item{chip_stacks}{Number of chips for each player}

\item{betting_units}{Size of betting unit for each player}

\item{minimum_bet}{Smallest bet allowed on the table}

\item{num_decks}{Number of decks of cards to use}
}
\value{
An object that represents all information in the game
}
\description{
Initialize a full table of players, including the dealer, and a random deck
}
