% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{create_player}
\alias{create_player}
\title{Create a Player}
\usage{
create_player(num_chips, betting_unit)
}
\arguments{
\item{num_chips}{Amount of money available to bet}
}
\value{
A list with information about the player
}
\description{
Initialize a player with a certain number of chips
}
