% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{create_players}
\alias{create_players}
\title{Create a table of players}
\usage{
create_players(chip_stacks)
}
\arguments{
\item{chip_stacks}{Number of chips for each player}
}
\value{
A list, with the dealer as the final element
}
\description{
Initialize a full table of players, including the dealer
}
