% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{deal_initial_cards}
\alias{deal_initial_cards}
\title{Deal Initial Cards to all Players}
\usage{
deal_initial_cards(game)
}
\arguments{
\item{game}{An object that represents all information in the game}
}
\value{
A game object
}
\description{
Deal Initial Cards to all Players
}
