% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{deal_new_card}
\alias{deal_new_card}
\title{Deal a New Card to a Player's Hand}
\usage{
deal_new_card(hand, new_card)
}
\arguments{
\item{hand}{A list of cards}

\item{new_card}{A new card}
}
\value{
An updated list of new cards
}
\description{
Deal a New Card to a Player's Hand
}
