% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deal_player.R
\name{Deal_player}
\alias{Deal_player}
\title{Deal the player their hand}
\usage{
Deal_player()
}
\value{
A list of two cards for the player
}
\description{
Deal the player their hand
}
\examples{
Deal_player()
}
