% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{dealer_has_blackjack}
\alias{dealer_has_blackjack}
\title{Does the dealer have blackjack}
\usage{
dealer_has_blackjack(game)
}
\arguments{
\item{game}{An object that represents all information in the game}
}
\value{
Boolean
}
\description{
Does the dealer have blackjack
}
