% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dealers_first.R
\name{Dealers_First}
\alias{Dealers_First}
\title{Show Dealer's card}
\usage{
Dealers_First(Deck)
}
\value{
The dealer's first card
}
\description{
Show Dealer's card
}
\examples{
Dealers_first(Deck1)
Dealers_first(Build_playing_deck(3))
}
