% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dealers_hand.R
\name{Dealers_hand}
\alias{Dealers_hand}
\title{Dealer's hand}
\usage{
Dealers_hand(Deck, Dealer)
}
\arguments{
\item{Deck}{Input the current deck in use.}

\item{Dealer}{Input the Dealer's up card}
}
\value{
Flips over the hole card, and displays both hole card and up card.
}
\description{
Dealer's hand
}
\examples{
Dealers_hand(Deck2, Up_card1)
}
