% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck_of_cards.R
\name{deck_cards}
\alias{deck_cards}
\title{Create a Blackjack Deck}
\usage{
deck_cards(no_of_decks = 1)
}
\arguments{
\item{no_of_decks}{Number of standard decks to include (default is 1)}
}
\value{
An object of class blackjack_deck, which is a list containing the shuffled cards and number of decks.
}
\description{
Generates a shuffled deck used in Blackjack.
You can specify the number of standard 52-card decks to include.
}
\examples{
deck <- deck_cards()
}
