% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{determine_bet_size}
\alias{determine_bet_size}
\title{How much should a given player bet}
\usage{
determine_bet_size(game, player_num)
}
\arguments{
\item{game}{An object that represents all information in the game}

\item{player_num}{The index of the player}
}
\value{
A number
}
\description{
How much should a given player bet
}
