% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{determine_hand_outcome}
\alias{determine_hand_outcome}
\title{Determine the outcome of the round}
\usage{
determine_hand_outcome(player_hand, dealer_hand)
}
\arguments{
\item{player_hand}{The player's cards}

\item{dealer_hand}{The dealer's cards}
}
\value{
One of "blackjack", "win", "push", or "lose"
}
\description{
Determine the outcome of the round
}
