% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{determine_player_action}
\alias{determine_player_action}
\title{Determine what decision a player should make when dealt cards}
\usage{
determine_player_action(player_hand, dealer_hand)
}
\arguments{
\item{player_hand}{A list of cards for a single hand}

\item{dealer_hand}{A list of cards}
}
\value{
One of 'hit', 'stand', 'double', 'split', or 'surrender'
}
\description{
Determine what decision a player should make when dealt cards
}
