% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{determine_soft_total_strategy}
\alias{determine_soft_total_strategy}
\title{Determine what decision a player should make on a soft total}
\usage{
determine_soft_total_strategy(player_total, dealer_total)
}
\arguments{
\item{player_total}{The player's total}

\item{dealer_total}{The dealer's visible total}
}
\value{
One of 'hit', 'stand', 'double', 'split', or 'surrender'
}
\description{
Determine what decision a player should make on a soft total
}
