% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{do_dealer_actions}
\alias{do_dealer_actions}
\title{Take Actions For Dealer}
\usage{
do_dealer_actions(game, verbose = 0)
}
\arguments{
\item{game}{An object that represents all information in the game}

\item{verbose}{An intger that determines how much information is print during
a simulation. 0 means no information. A higher integer results in more printing.}
}
\value{
An updated game object
}
\description{
Take Actions For Dealer
}
