% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackjack_functions.R
\name{doubleDown}
\alias{doubleDown}
\title{doubleDown}
\usage{
doubleDown(player_total, cards_remaining, dealer_card, is_ace)
}
\arguments{
\item{player_total}{sum of the two cards so far}

\item{cards_remaining}{vector with number of each type of card remaining}

\item{dealer_card}{dealer card}

\item{is_ace}{\code{TRUE} if the hand contains an ace}
}
\value{
expected winnings of the double down
}
\description{
returns unit ROI for a double down (hit exactly one more card)
}
\author{
David Clement
}
