% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackjack_functions.R
\name{findAllCards}
\alias{findAllCards}
\title{findAllCards}
\usage{
findAllCards(n_decks)
}
\arguments{
\item{n_decks}{number of decks in play}
}
\value{
a named vector of length 10 with card counts at start of game
}
\description{
find a vector of length 10 with card counts at start of game
}
\author{
David Clement
}
