% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackjack_functions.R
\name{findRemainingCards}
\alias{findRemainingCards}
\title{findRemainingCards}
\usage{
findRemainingCards(cards_gone, n_decks)
}
\arguments{
\item{cards_gone}{vector of cards gone, e.g. c(1,10,5) would be one ace,
one 10/face card, and one 5 already showing}

\item{n_decks}{number of decks in play}
}
\value{
a named vector of length 10 with card counts at the current state
of the game
}
\description{
vector of length 10 with counts of remaining cards
}
\author{
David Clement
}
