% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{get_best_hand_total}
\alias{get_best_hand_total}
\title{Get Hand Total}
\usage{
get_best_hand_total(hand)
}
\arguments{
\item{hand}{A list of cards}
}
\value{
The best numeric value that a hand can make without going over
}
\description{
Determine the total value of the cards in a hand
}
