% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{get_hand_total}
\alias{get_hand_total}
\title{Get Hand Total}
\usage{
get_hand_total(hand)
}
\arguments{
\item{hand}{A list of cards}
}
\value{
A list containing the numeric total and an indicator on whether the
hand includes an ace
}
\description{
Determine the total value of the cards in a hand
}
