% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{is_black_jack}
\alias{is_black_jack}
\title{Determine if a hand is a 2-card 21}
\usage{
is_black_jack(hand)
}
\arguments{
\item{hand}{A list of cards}
}
\value{
Boolean
}
\description{
Determine if a hand is a 2-card 21
}
