% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{is_hand_a_pair}
\alias{is_hand_a_pair}
\title{Determine if a players hand is a pair}
\usage{
is_hand_a_pair(hand)
}
\arguments{
\item{hand}{A list of cards}
}
\value{
Boolean
}
\description{
Determine if a players hand is a pair
}
