% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{is_hand_a_soft_ace}
\alias{is_hand_a_soft_ace}
\title{Determine if a player's hand has a soft ace}
\usage{
is_hand_a_soft_ace(hand_total)
}
\arguments{
\item{hand_total}{A list with the total value of the ranks and a flag on
whether the hand has an ace. The first ace in the hand contributes 1 to the total.}
}
\value{
Boolean
}
\description{
Determine if a player's hand has a soft ace
}
