% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{is_hand_active}
\alias{is_hand_active}
\title{Is this hand not busted or black jack}
\usage{
is_hand_active(hand)
}
\arguments{
\item{hand}{A list of cards}
}
\value{
Boolean
}
\description{
Is this hand not busted or black jack
}
