% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{is_hand_busted}
\alias{is_hand_busted}
\title{Determine if a hand has gone over 21}
\usage{
is_hand_busted(hand)
}
\arguments{
\item{hand}{A list of cards}
}
\value{
Boolean
}
\description{
Determine if a hand has gone over 21
}
