% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{is_one_player_still_active}
\alias{is_one_player_still_active}
\title{Are any hands not busted or black jack}
\usage{
is_one_player_still_active(game)
}
\arguments{
\item{game}{An object that represents all information in the game}
}
\value{
Boolean
}
\description{
Are any hands not busted or black jack
}
