% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyHand.R
\name{MyHand}
\alias{MyHand}
\title{Deal the player their hand}
\usage{
MyHand(Deck)
}
\arguments{
\item{Deck}{Enter the current deck within the function}
}
\value{
Returns two cards for the player's original hand.
}
\description{
Deal the player their hand
}
\examples{
MyHand(Build_playing_deck(2))
MyHand(Deck2)
}
