% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nextCardProbs}
\alias{nextCardProbs}
\title{The probabilities of each card coming next}
\usage{
nextCardProbs(cards_remaining, dealer_card)
}
\arguments{
\item{cards_remaining}{vector count of each type of card}

\item{dealer_card}{value of dealer's card showing}
}
\value{
vector of probabilities
}
\description{
The probabilities of each card coming next
}
