% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_nonrcpp_functions.R
\name{nextCardProbs2}
\alias{nextCardProbs2}
\title{nextCardProbs2}
\usage{
nextCardProbs2(cards_remaining, dealer_card = NA)
}
\arguments{
\item{cards_remaining}{vector with number of each type of card remaining}

\item{dealer_card}{the card the dealer has. if 1 then they can't get 10 next,
if 10 then they can't get 1 next. NA is treated like 2:9}
}
\value{
vector of adjusted player next card probabilities
}
\description{
not only do we have to adjust the \code{dealer} function when he's
showing 1 or 10, we have to slightly modify the player next card probs
}
\author{
David Clement
}
