% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{place_bets}
\alias{place_bets}
\title{Move Money From Chip Stacks To Place Bets}
\usage{
place_bets(game)
}
\arguments{
\item{game}{An object that represents all information in the game}
}
\value{
A game object
}
\description{
Move Money From Chip Stacks To Place Bets
}
