% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{play_round}
\alias{play_round}
\title{Play a round a blackjack}
\usage{
play_round(game)
}
\arguments{
\item{game}{An object that represents all information in the game}
}
\value{
An updated game object
}
\description{
Play a round a blackjack
}
