% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{play_rounds}
\alias{play_rounds}
\title{Play multiple rounds of blackjack}
\usage{
play_rounds(
  chip_stacks,
  betting_units,
  minimum_bet,
  num_decks,
  num_rounds,
  verbose = 0
)
}
\arguments{
\item{chip_stacks}{Number of chips for each player}

\item{betting_units}{Size of betting unit for each player}

\item{minimum_bet}{Smallest bet allowed on the table}

\item{num_decks}{Number of decks of cards to use}

\item{num_rounds}{An integer representing the number of rounds to play}

\item{verbose}{An intger that determines how much information is print during
a simulation. 0 means no information. A higher integer results in more printing.}
}
\value{
An object that represents all information in the game
}
\description{
Play multiple rounds of blackjack
}
