% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_chip_counts}
\alias{plot_chip_counts}
\title{Plot Chip Counts for Each Player Over Time}
\usage{
plot_chip_counts(chip_counts)
}
\arguments{
\item{chip_counts}{A matrix, where each row is a round and each column is a player}
}
\value{
A ggplot object
}
\description{
Plot Chip Counts for Each Player Over Time
}
