% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{record_chip_counts}
\alias{record_chip_counts}
\title{Add Player Chip Counts to a Vector}
\usage{
record_chip_counts(game)
}
\arguments{
\item{game}{An object that represents all information in the game}
}
\value{
A vector where each element is the number of chips for a given player
}
\description{
Add Player Chip Counts to a Vector
}
