% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{reset_round}
\alias{reset_round}
\title{Reset Chip Stacks and Hands}
\usage{
reset_round(game)
}
\arguments{
\item{game}{An object that represents all information in the game}
}
\value{
An updated game object
}
\description{
Reset Chip Stacks and Hands
}
