% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blackjack.R
\name{run_blackjack}
\alias{run_blackjack}
\title{Lancer l'application Shiny Blackjack}
\usage{
run_blackjack()
}
\value{
Lance une application Shiny
}
\description{
Cette fonction lance l'application Shiny incluse dans le package blackjack.
}
