% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{should_dealer_hit}
\alias{should_dealer_hit}
\title{Should Dealer Hit}
\usage{
should_dealer_hit(hand_total)
}
\arguments{
\item{hand_total}{A hand total}
}
\value{
Boolean
}
\description{
Dealer hits on a soft 16
}
