% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{should_player_split_pairs}
\alias{should_player_split_pairs}
\title{Determine if a player should split a pair}
\usage{
should_player_split_pairs(player_total, dealer_total)
}
\arguments{
\item{player_total}{The player's total}

\item{dealer_total}{The dealer's visible total}
}
\value{
Boolean
}
\description{
Determine if a player should split a pair
}
