% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shuffle.R
\name{Shuffle}
\alias{Shuffle}
\title{Shuffle Function}
\usage{
Shuffle(Deck)
}
\arguments{
\item{Deck}{Input the deck of cards you wish too be shuffled}
}
\value{
Will return a function capable of randomizing the deck
}
\description{
Shuffle Function
}
\examples{
Shuffle(Deck2)
}
