% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{shuffle_cards}
\alias{shuffle_cards}
\title{Shuffle a Deck of Cards}
\usage{
shuffle_cards(num_decks)
}
\arguments{
\item{num_decks}{The number of decks you want to include in the shuffle.}
}
\value{
A list, where each element represents a card.
}
\description{
Uses a standard 52 card deck. A card is represented as a list with two
attributes: rank and suit. A deck is a list of cards.
}
