% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackjack_functions.R
\name{split}
\alias{split}
\title{split}
\usage{
split(p, cards_remaining, dealer_card, blackjack_payout, n_hands = 2)
}
\arguments{
\item{p}{the value of the card being split}

\item{cards_remaining}{vector with number of each type of card remaining}

\item{dealer_card}{dealer card}

\item{blackjack_payout}{this is the one place in hit/stand/split/double down
decision process (without card counting) where blackjack odds are needed}

\item{n_hands}{I think this function can be written faster without this argument
but as a first pass I found it clearer this way. note also I am allowing for a
situation where you split once but then don't want to split again - I would
theorize this would never actually happen. You wouldn't split the first time.}
}
\value{
expected winnings of the split
}
\description{
returns unit ROI for a split
}
\author{
David Clement
}
