% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackjack_functions.R
\name{stand}
\alias{stand}
\title{stand}
\usage{
stand(player_total, dealer_total, cards_remaining)
}
\arguments{
\item{player_total}{sum of player cards so far in the hand}

\item{dealer_total}{what dealer card is showing}

\item{cards_remaining}{vector of counts for each of 1 through 10}
}
\value{
a player win probability
}
\description{
function to find player win probability if they stand
}
\author{
David Clement
}
