% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Up_card.R
\name{Up_card}
\alias{Up_card}
\title{Show Dealer's card}
\usage{
Up_card(Deck)
}
\arguments{
\item{Deck}{The deck currently being used to play}
}
\value{
The dealer's first card
}
\description{
Show Dealer's card
}
\examples{
Dealers_first(Deck1)
Dealers_first(Build_playing_deck(3))
}
