% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{update_chip_stacks}
\alias{update_chip_stacks}
\title{Give Additional Chips Based On Round Outcomes}
\usage{
update_chip_stacks(game, verbose = 0)
}
\arguments{
\item{game}{An object that represents all information in the game}

\item{verbose}{An intger that determines how much information is print during
a simulation. 0 means no information. A higher integer results in more printing.}
}
\value{
An updated game object
}
\description{
Give Additional Chips Based On Round Outcomes
}
