% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{update_count}
\alias{update_count}
\title{Update value of the count based on a new card}
\usage{
update_count(game, card)
}
\arguments{
\item{game}{An object that represents all information in the game}

\item{card}{A list containing two attributes: rank and suit}
}
\value{
A game object
}
\description{
Update value of the count based on a new card
}
