% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{create_player}
\alias{create_player}
\title{Create a Player}
\usage{
create_player(num_chips)
}
\arguments{
\item{num_chips}{Amount of money available to bet}
}
\value{
A list with the players role, number of chips, and active hands
}
\description{
Initialize a player with a certain number of chips
}
