% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dealer}
\alias{dealer}
\title{vector of probabilities for the 6 different dealer possibilities}
\usage{
dealer(dealer_total, cards_remaining, is_ace, hit_soft_17, is_first)
}
\arguments{
\item{dealer_total}{dealer's total so far}

\item{cards_remaining}{vector count of each type of card}

\item{is_ace}{boolean indicating whether hand is soft}

\item{hit_soft_17}{int (0 or 1) indicating dealer behaviour; defaults to true}

\item{is_first}{boolean indicating whether this is dealer's first unknown card}
}
\value{
vector of probabilities (length 6)
}
\description{
vector of probabilities for the 6 different dealer possibilities
}
