% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{determine_bet_size}
\alias{determine_bet_size}
\title{Determine what amount to bet at the beginning of the hand}
\usage{
determine_bet_size(game)
}
\arguments{
\item{game}{An object that represents all information in the game}
}
\value{
A numeric value
}
\description{
Determine what amount to bet at the beginning of the hand
}
