% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{do_dealer_actions}
\alias{do_dealer_actions}
\title{Take Actions For Dealer}
\usage{
do_dealer_actions(game)
}
\arguments{
\item{game}{An object that represents all information in the game}
}
\value{
An updated game object
}
\description{
Take Actions For Dealer
}
