% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hit.R
\name{Hit}
\alias{Hit}
\title{The "Hit" function}
\usage{
Hit(Deck, hand)
}
\arguments{
\item{Deck}{Input the deck currently in play}

\item{hand}{Input the hand that needs a new card}
}
\description{
The "Hit" function
}
\examples{
Hit(Deck1,Hand1)
Hit(Deck2, DealerHand)
}
