% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackjack_functions.R
\name{hit}
\alias{hit}
\title{hit}
\usage{
hit(player_total, dealer_total, cards_remaining)
}
\arguments{
\item{player_total}{sum of player cards so far in the hand}

\item{dealer_total}{what dealer card is showing}

\item{cards_remaining}{vector of counts for each of 1 through 10}
}
\value{
a player win probability
}
\description{
recursive function to find player win probability when they hit
}
\author{
David Clement
}
