% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackjack_functions.R
\name{hitStandProbs}
\alias{hitStandProbs}
\title{hitStandProbs}
\usage{
hitStandProbs(p1, p2, d, cards_gone = c(), n_decks = 1)
}
\arguments{
\item{p1}{first player card (order with p2 doesn't matter)}

\item{p2}{second player card}

\item{d}{dealer card}

\item{cards_gone}{will be modified to include p1, p2, d. But could also
pass initial vector here too for card counting purposes}

\item{n_decks}{default to 1 while I test (for efficiency purposes)}
}
\value{
data frame with p_win_if_stand, p_win_if_hit, player_card_1,
player_card_2, dealer_card
}
\description{
compares the hand win probability for hit and stand given the
player and dealer starting cards -- test commit does this work
}
\author{
David Clement
}
