% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{play_round}
\alias{play_round}
\title{Play a round a blackjack}
\usage{
play_round(game, verbose = 0)
}
\arguments{
\item{game}{An object that represents all information in the game}

\item{verbose}{An intger that determines how much information is print during
a simulation. 0 means no information. A higher integer results in more printing.}
}
\value{
An updated game object
}
\description{
Play a round a blackjack
}
