% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shuffle.R
\name{Shuffle}
\alias{Shuffle}
\title{Shuffle Function}
\usage{
Shuffle(Deck)
}
\arguments{
\item{Deck}{Enter the deck we want to shuffle up}
}
\value{
Returns a well-mixed deck
}
\description{
Shuffle Function
}
\examples{
Shuffle(Deck1)
}
