% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{update_chip_stacks}
\alias{update_chip_stacks}
\title{Give Additional Chips Based On Round Outcomes}
\usage{
update_chip_stacks(game)
}
\arguments{
\item{game}{An object that represents all information in the game}
}
\value{
An updated game object
}
\description{
Give Additional Chips Based On Round Outcomes
}
